AG-SEIIHURD fit
First version of fit the age group SEIIHURD model, using data from SESAB, reorganized with 2 groups (above and below 60 years-old) 

The scipy least_squares with multiple random starting works as intended and finds a reasonable 'solution'. 
The PSO method does not converge with the parameters employed. It may be a bug in the objective function or the method is not configured correctly.