Bug Fixes
Fixed a bug that made the bounds of betas be wrongly defined when the flag "_ALL" was used. 

Small code changes, now there is a prepare_to_fit function that allows the use of objectiveFunction (and the bounds) with external minimizers.

An example of using this code to fit a data series with only one age group.